+++ homeassistant/components/conversation/models.py
